+++ src/daf_gui/tabs/__init__.py
@@ -1,2 +1,3 @@
 from .schema import *
 from .analytics import *
+from .optional import *
